+++ Bliss/Manager/SizeManager.cs
@@ -1,8 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
-using System;
-using System.Collections.Generic;
 using System.Drawing;
-using System.Text;
 using Unity;
 
 namespace Bliss.Manager
@@ -12,9 +9,9 @@ namespace Bliss.Manager
         [Dependency]
         public JamGame JamGame { get; set; }
 
-        public Size GetSize(int width, int height) => new Size(width * (int)JamGame.WidthScaleFactor, height * (int)JamGame.HeightScaleFactor);
-        public Vector2 GetPosition(int x, int y) => new Vector2(x * (int)JamGame.WidthScaleFactor, y * (int)JamGame.HeightScaleFactor);
-        public float ScaleForWidth(int width) => width * (int)JamGame.WidthScaleFactor;
-        public float ScaleForHeight(int height) => height * (int)JamGame.HeightScaleFactor;
+        public Size GetSize(int width, int height) => new Size((int)(width * JamGame.WidthScaleFactor), (int)(height * JamGame.HeightScaleFactor));
+        public Vector2 GetPosition(int x, int y) => new Vector2(x * JamGame.WidthScaleFactor, y * JamGame.HeightScaleFactor);
+        public float ScaleForWidth(int width) => width * JamGame.WidthScaleFactor;
+        public float ScaleForHeight(int height) => height * JamGame.HeightScaleFactor;
     }
 }
